[hal] Move registerSimPeriodic functions to HAL package
This enables the static lists to be private.
