+++ hal/src/main/java/edu/wpi/first/hal/HAL.java
@@ -28,7 +28,36 @@ public final class HAL extends JNIWrapper {
 
   private static native void simPeriodicBeforeNative();
 
-  public static final List<Runnable> s_simPeriodicBefore = new ArrayList<>();
+  private static final List<Runnable> s_simPeriodicBefore = new ArrayList<>();
+
+  public static class SimPeriodicBeforeCallback implements AutoCloseable {
+    private SimPeriodicBeforeCallback(Runnable r) {
+      m_run = r;
+    }
+
+    @Override
+    public void close() {
+      synchronized (s_simPeriodicBefore) {
+        s_simPeriodicBefore.remove(m_run);
+      }
+    }
+
+    private final Runnable m_run;
+  }
+
+  /**
+   * Registers a callback to be run by IterativeRobotBase prior to the user's simulationPeriodic
+   * code.
+   *
+   * @param r runnable
+   * @return Callback object (must be retained for callback to stay active).
+   */
+  public static SimPeriodicBeforeCallback registerSimPeriodicBeforeCallback(Runnable r) {
+    synchronized (s_simPeriodicBefore) {
+      s_simPeriodicBefore.add(r);
+    }
+    return new SimPeriodicBeforeCallback(r);
+  }
 
   /**
    * Runs SimPeriodicBefore callbacks. IterativeRobotBase calls this prior to the user's
@@ -45,7 +74,35 @@ public final class HAL extends JNIWrapper {
 
   private static native void simPeriodicAfterNative();
 
-  public static final List<Runnable> s_simPeriodicAfter = new ArrayList<>();
+  private static final List<Runnable> s_simPeriodicAfter = new ArrayList<>();
+
+  public static class SimPeriodicAfterCallback implements AutoCloseable {
+    private SimPeriodicAfterCallback(Runnable r) {
+      m_run = r;
+    }
+
+    @Override
+    public void close() {
+      synchronized (s_simPeriodicAfter) {
+        s_simPeriodicAfter.remove(m_run);
+      }
+    }
+
+    private final Runnable m_run;
+  }
+
+  /**
+   * Registers a callback to be run by IterativeRobotBase after the user's simulationPeriodic code.
+   *
+   * @param r runnable
+   * @return Callback object (must be retained for callback to stay active).
+   */
+  public static SimPeriodicAfterCallback registerSimPeriodicAfterCallback(Runnable r) {
+    synchronized (s_simPeriodicAfter) {
+      s_simPeriodicAfter.add(r);
+    }
+    return new SimPeriodicAfterCallback(r);
+  }
 
   /**
    * Runs SimPeriodicAfter callbacks. IterativeRobotBase calls this after the user's
